+++ CollegeGradingSys/Models/Repositories/NationalityDbRepository.cs
@@ -11,53 +11,50 @@ namespace CollegeGradingSys.Models.Repositories
         private readonly ApplicationDbContext db;
 
         public NationalityDbRepository(ApplicationDbContext _db)
-    {
+        {
             db = _db;
         }
-
-
-
-
-
-
-
-
-
-
-    public void Add(Nationality entity)
-    {  
-        db.Nationality.Add(entity);
+        public Nationality Add(Nationality entity)
+        {
+            db.Nationality.Add(entity);
             SaveChange();
-    }
+            return entity;
+        }
 
-    public void Delete(int id)
-    {
-        var nationality = Find(id);
-            db.Nationality.Remove(nationality);
+        public Nationality Delete(int id)
+        {
+            var nationality = Find(id);
+            if (nationality != null)
+            {
+                db.Nationality.Remove(nationality);
+                SaveChange();
+            }
+            return nationality;
+        }
+
+        public Nationality Find(int id)
+        {
+            return db.Nationality.SingleOrDefault(a => a.Id == id);
+        }
+
+        public IList<Nationality> List()
+        {
+            return db.Nationality.ToList();
+        }
+
+        public Nationality Update(int id, Nationality newNationality)
+        {
+            var college = db.Nationality.Attach(newNationality);
+            college.State = Microsoft.EntityFrameworkCore.EntityState.Modified;
             SaveChange();
-    }
-
-    public Nationality Find(int id)
-    {
-        return db.Nationality.SingleOrDefault(a => a.Id == id);
-    }
-
-    public IList<Nationality> List()
-    {
-        return db.Nationality.ToList();
-    }
-
-    public void Update(int id, Nationality newNationality)
-    {
-            db.Nationality.Update(newNationality);
-            SaveChange();
-    }
+            return newNationality;
+        }
 
         private void SaveChange()
         {
             db.SaveChanges();
         }
-}
+    }
 }
